@@ -1,12 +1,12 @@
 [tool.poetry]
 name = "safe-control-gym"
-version = "0.5.0"
+version = "2.0.0"
 description = "PyBullet CartPole and Quadrotor environments—with CasADi symbolic a priori dynamics—for learning-based control and RL"
 authors = ["Your Name <you@example.com>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.8.10"
+python = "^3.10"
 matplotlib = "^3.5.1"
 Pillow = "^9.0.0"
 munch = "^2.5.0"
@@ -18,14 +18,13 @@ pandas = "^1.4.0"
 gymnasium = "^0.28"
 torch = "^1.10.2"
 gpytorch = "^1.6.0"
-ray = "^1.9.2"
-tensorboard = "^2.8.0"
-casadi = "3.5.5"
+tensorboard = "^2.12.0"
+casadi = "^3.6.0"
 pybullet = "^3.2.0"
-cvxpy = "^1.1.18"
 numpy = "^1.22.1"
+cvxpy = "^1.1.18"
 pytope = "^0.0.4"
-Mosek = "^9.3.14"
+Mosek = "^10.0.18"
 termcolor = "^1.1.0"
 pytest = "^7.2.2"
 
