@@ -1,4 +1,4 @@
-"""Deep Deterministic Policy Gradient
+'''Deep Deterministic Policy Gradient
 
 Reference paper & code:
     * [Continuous Control with Deep Reinforcement Learning](https://arxiv.org/pdf/1509.02971.pdf)
@@ -7,13 +7,14 @@ Reference paper & code:
 
 Example:
     $ python safe_control_gym/experiments/execute_rl_controller.py --algo ddpg --task cartpole --output_dir results --tag test/cartpole_ddpg --seed 6
-"""
+'''
 
 import os
 import time
+from collections import defaultdict
+
 import numpy as np
 import torch
-from collections import defaultdict
 
 from safe_control_gym.utils.logging import ExperimentLogger
 from safe_control_gym.utils.utils import get_random_state, set_random_state, is_wrapped
@@ -27,13 +28,13 @@ from safe_control_gym.controllers.ddpg.ddpg_utils import DDPGAgent, DDPGBuffer, 
 
 
 class DDPG(BaseController):
-    """deep deterministic policy gradient."""
+    '''Deep Deterministic Policy Gradient.'''
 
     def __init__(self,
                  env_func,
                  training=True,
-                 checkpoint_path="model_latest.pt",
-                 output_dir="temp",
+                 checkpoint_path='model_latest.pt',
+                 output_dir='temp',
                  use_gpu=False,
                  seed=0,
                  **kwargs):
@@ -57,12 +58,12 @@ class DDPG(BaseController):
 
         # agent
         self.agent = DDPGAgent(self.env.observation_space,
-                              self.env.action_space,
-                              hidden_dim=self.hidden_dim,
-                              gamma=self.gamma,
-                              tau=self.tau,
-                              actor_lr=self.actor_lr,
-                              critic_lr=self.critic_lr)
+                               self.env.action_space,
+                               hidden_dim=self.hidden_dim,
+                               gamma=self.gamma,
+                               tau=self.tau,
+                               actor_lr=self.actor_lr,
+                               critic_lr=self.critic_lr)
         self.agent.to(self.device)
 
         # pre-/post-processing
@@ -85,13 +86,13 @@ class DDPG(BaseController):
         self.logger = ExperimentLogger(output_dir, log_file_out=log_file_out, use_tensorboard=use_tensorboard)
 
     def reset(self):
-        """Prepares for training or testing."""
+        '''Prepares for training or testing.'''
         if self.training:
             # set up stats tracking
-            self.env.add_tracker("constraint_violation", 0)
-            self.env.add_tracker("constraint_violation", 0, mode="queue")
-            self.eval_env.add_tracker("constraint_violation", 0, mode="queue")
-            self.eval_env.add_tracker("mse", 0, mode="queue")
+            self.env.add_tracker('constraint_violation', 0)
+            self.env.add_tracker('constraint_violation', 0, mode='queue')
+            self.eval_env.add_tracker('constraint_violation', 0, mode='queue')
+            self.eval_env.add_tracker('mse', 0, mode='queue')
 
             self.total_steps = 0
             obs, _ = self.env.reset()
@@ -102,67 +103,67 @@ class DDPG(BaseController):
                 self.noise_process.reset_states()
         else:
             # set up stats tracking
-            self.env.add_tracker("constraint_violation", 0, mode="queue")
-            self.env.add_tracker("constraint_values", 0, mode="queue")
-            self.env.add_tracker("mse", 0, mode="queue")
+            self.env.add_tracker('constraint_violation', 0, mode='queue')
+            self.env.add_tracker('constraint_values', 0, mode='queue')
+            self.env.add_tracker('mse', 0, mode='queue')
 
     def close(self):
-        """Shuts down and cleans up lingering resources."""
+        '''Shuts down and cleans up lingering resources.'''
         self.env.close()
         if self.training:
             self.eval_env.close()
         self.logger.close()
 
     def save(self, path, save_buffer=True):
-        """Saves model params and experiment state to checkpoint path."""
+        '''Saves model params and experiment state to checkpoint path.'''
         path_dir = os.path.dirname(path)
         os.makedirs(path_dir, exist_ok=True)
 
         state_dict = {
-            "agent": self.agent.state_dict(),
-            "obs_normalizer": self.obs_normalizer.state_dict(),
-            "reward_normalizer": self.reward_normalizer.state_dict()
+            'agent': self.agent.state_dict(),
+            'obs_normalizer': self.obs_normalizer.state_dict(),
+            'reward_normalizer': self.reward_normalizer.state_dict()
         }
         if self.training:
             exp_state = {
-                "total_steps": self.total_steps,
-                "obs": self.obs,
-                "random_state": get_random_state(),
-                "env_random_state": self.env.get_env_random_state()
+                'total_steps': self.total_steps,
+                'obs': self.obs,
+                'random_state': get_random_state(),
+                'env_random_state': self.env.get_env_random_state()
             }
             # latest checkpoint shoud enable save_buffer (for experiment restore),
             # but intermediate checkpoint shoud not, to save storage (buffer is large)
             if save_buffer:
-                exp_state["buffer"] = self.buffer.state_dict()
+                exp_state['buffer'] = self.buffer.state_dict()
             # noise process is also stateful
             if self.noise_process:
-                exp_state["noise_process"] = self.noise_process.state_dict()
+                exp_state['noise_process'] = self.noise_process.state_dict()
             state_dict.update(exp_state)
         torch.save(state_dict, path)
 
     def load(self, path):
-        """Restores model and experiment given checkpoint path."""
+        '''Restores model and experiment given checkpoint path.'''
         state = torch.load(path)
 
         # restore params
-        self.agent.load_state_dict(state["agent"])
-        self.obs_normalizer.load_state_dict(state["obs_normalizer"])
-        self.reward_normalizer.load_state_dict(state["reward_normalizer"])
+        self.agent.load_state_dict(state['agent'])
+        self.obs_normalizer.load_state_dict(state['obs_normalizer'])
+        self.reward_normalizer.load_state_dict(state['reward_normalizer'])
 
         # restore experiment state
         if self.training:
-            self.total_steps = state["total_steps"]
-            self.obs = state["obs"]
-            set_random_state(state["random_state"])
-            self.env.set_env_random_state(state["env_random_state"])
-            if "buffer" in state:
-                self.buffer.load_state_dict(state["buffer"])
+            self.total_steps = state['total_steps']
+            self.obs = state['obs']
+            set_random_state(state['random_state'])
+            self.env.set_env_random_state(state['env_random_state'])
+            if 'buffer' in state:
+                self.buffer.load_state_dict(state['buffer'])
             if self.noise_process:
-                self.noise_process.load_state_dict(state["noise_process"])
+                self.noise_process.load_state_dict(state['noise_process'])
             self.logger.load(self.total_steps)
 
     def learn(self, env=None, **kwargs):
-        """Performs learning (pre-training, training, fine-tuning, etc)."""
+        '''Performs learning (pre-training, training, fine-tuning, etc).'''
         while self.total_steps < self.max_env_steps:
             results = self.train_step()
 
@@ -170,33 +171,33 @@ class DDPG(BaseController):
             if self.total_steps >= self.max_env_steps or (self.save_interval and self.total_steps % self.save_interval == 0):
                 # latest/final checkpoint
                 self.save(self.checkpoint_path)
-                self.logger.info("Checkpoint | {}".format(self.checkpoint_path))
+                self.logger.info(f'Checkpoint | {self.checkpoint_path}')
             if self.num_checkpoints and self.total_steps % (self.max_env_steps // self.num_checkpoints) == 0:
                 # intermediate checkpoint
-                path = os.path.join(self.output_dir, "checkpoints", "model_{}.pt".format(self.total_steps))
+                path = os.path.join(self.output_dir, 'checkpoints', f'model_{self.total_steps}.pt')
                 self.save(path, save_buffer=False)
 
             # eval
             if self.eval_interval and self.total_steps % self.eval_interval == 0:
                 eval_results = self.run(env=self.eval_env, n_episodes=self.eval_batch_size)
-                results["eval"] = eval_results
-                self.logger.info("Eval | ep_lengths {:.2f} +/- {:.2f} | ep_return {:.3f} +/- {:.3f}".format(eval_results["ep_lengths"].mean(),
-                                                                                                            eval_results["ep_lengths"].std(),
-                                                                                                            eval_results["ep_returns"].mean(),
-                                                                                                            eval_results["ep_returns"].std()))
+                results['eval'] = eval_results
+                self.logger.info('Eval | ep_lengths {:.2f} +/- {:.2f} | ep_return {:.3f} +/- {:.3f}'.format(eval_results['ep_lengths'].mean(),
+                                                                                                            eval_results['ep_lengths'].std(),
+                                                                                                            eval_results['ep_returns'].mean(),
+                                                                                                            eval_results['ep_returns'].std()))
                 # save best model
-                eval_score = eval_results["ep_returns"].mean()
-                eval_best_score = getattr(self, "eval_best_score", -np.infty)
+                eval_score = eval_results['ep_returns'].mean()
+                eval_best_score = getattr(self, 'eval_best_score', -np.infty)
                 if self.eval_save_best and eval_best_score < eval_score:
                     self.eval_best_score = eval_score
-                    self.save(os.path.join(self.output_dir, "model_best.pt"))
+                    self.save(os.path.join(self.output_dir, 'model_best.pt'))
 
             # logging
             if self.log_interval and self.total_steps % self.log_interval == 0:
                 self.log_step(results)
 
     def select_action(self, obs, info=None):
-        """Determine the action to take at the current timestep.
+        '''Determine the action to take at the current timestep.
 
         Args:
             obs (ndarray): The observation at this timestep.
@@ -204,7 +205,7 @@ class DDPG(BaseController):
 
         Returns:
             action (ndarray): The action chosen by the controller.
-        """
+        '''
 
         with torch.no_grad():
             obs = torch.FloatTensor(obs).to(self.device)
@@ -213,7 +214,7 @@ class DDPG(BaseController):
         return action
 
     def run(self, env=None, render=False, n_episodes=10, verbose=False, **kwargs):
-        """Runs evaluation with current policy."""
+        '''Runs evaluation with current policy.'''
         self.agent.eval()
         self.obs_normalizer.set_read_only()
         if env is None:
@@ -222,9 +223,9 @@ class DDPG(BaseController):
             if not is_wrapped(env, RecordEpisodeStatistics):
                 env = RecordEpisodeStatistics(env, n_episodes)
                 # Add episodic stats to be tracked.
-                env.add_tracker("constraint_violation", 0, mode="queue")
-                env.add_tracker("constraint_values", 0, mode="queue")
-                env.add_tracker("mse", 0, mode="queue")
+                env.add_tracker('constraint_violation', 0, mode='queue')
+                env.add_tracker('constraint_values', 0, mode='queue')
+                env.add_tracker('mse', 0, mode='queue')
 
         obs, info = env.reset()
         obs = self.obs_normalizer(obs)
@@ -234,26 +235,26 @@ class DDPG(BaseController):
         while len(ep_returns) < n_episodes:
             action = self.select_action(obs, info)
 
-            obs, reward, done, info = env.step(action)
+            obs, _, done, info = env.step(action)
             if render:
                 env.render()
-                frames.append(env.render("rgb_array"))
+                frames.append(env.render('rgb_array'))
             if verbose:
-                print("obs {} | act {}".format(obs, action))
+                print(f'obs {obs} | act {action}')
 
             if done:
-                assert "episode" in info
-                ep_returns.append(info["episode"]["r"])
-                ep_lengths.append(info["episode"]["l"])
+                assert 'episode' in info
+                ep_returns.append(info['episode']['r'])
+                ep_lengths.append(info['episode']['l'])
                 obs, info = env.reset()
             obs = self.obs_normalizer(obs)
 
         # collect evaluation results
         ep_lengths = np.asarray(ep_lengths)
         ep_returns = np.asarray(ep_returns)
-        eval_results = {"ep_returns": ep_returns, "ep_lengths": ep_lengths}
+        eval_results = {'ep_returns': ep_returns, 'ep_lengths': ep_lengths}
         if len(frames) > 0:
-            eval_results["frames"] = frames
+            eval_results['frames'] = frames
         # Other episodic stats from evaluation env.
         if len(env.queued_stats) > 0:
             queued_stats = {k: np.asarray(v) for k, v in env.queued_stats.items()}
@@ -261,7 +262,7 @@ class DDPG(BaseController):
         return eval_results
 
     def train_step(self, **kwargs):
-        """Performs a training step."""
+        '''Performs a training step.'''
         self.agent.train()
         self.obs_normalizer.unset_read_only()
         obs = self.obs
@@ -284,13 +285,13 @@ class DDPG(BaseController):
 
         # time truncation is not true termination
         terminal_idx, terminal_obs = [], []
-        for idx, inf in enumerate(info["n"]):
-            if "terminal_info" not in inf:
+        for idx, inf in enumerate(info['n']):
+            if 'terminal_info' not in inf:
                 continue
-            inff = inf["terminal_info"]
-            if "TimeLimit.truncated" in inff and inff["TimeLimit.truncated"]:
+            inff = inf['terminal_info']
+            if 'TimeLimit.truncated' in inff and inff['TimeLimit.truncated']:
                 terminal_idx.append(idx)
-                terminal_obs.append(inf["terminal_observation"])
+                terminal_obs.append(inf['terminal_observation'])
         if len(terminal_obs) > 0:
             terminal_obs = _unflatten_obs(self.obs_normalizer(_flatten_obs(terminal_obs)))
 
@@ -303,13 +304,13 @@ class DDPG(BaseController):
         true_next_obs = _flatten_obs(true_next_obs)
 
         self.buffer.push({
-            "obs": obs,
-            "act": act,
-            "rew": rew,
-            # "next_obs": next_obs,
-            # "mask": mask,
-            "next_obs": true_next_obs,
-            "mask": true_mask,
+            'obs': obs,
+            'act': act,
+            'rew': rew,
+            # 'next_obs': next_obs,
+            # 'mask': mask,
+            'next_obs': true_next_obs,
+            'mask': true_mask,
         })
         obs = next_obs
 
@@ -329,67 +330,67 @@ class DDPG(BaseController):
                     results[k].append(v)
 
         results = {k: sum(v) / len(v) for k, v in results.items()}
-        results.update({"step": self.total_steps, "elapsed_time": time.time() - start})
+        results.update({'step': self.total_steps, 'elapsed_time': time.time() - start})
         return results
 
     def log_step(self, results):
-        """Does logging after a training step."""
-        step = results["step"]
+        '''Does logging after a training step.'''
+        step = results['step']
         # runner stats
         self.logger.add_scalars(
             {
-                "step": step,
-                "time": results["elapsed_time"],
-                "progress": step / self.max_env_steps,
+                'step': step,
+                'time': results['elapsed_time'],
+                'progress': step / self.max_env_steps,
             },
             step,
-            prefix="time",
+            prefix='time',
             write=False,
             write_tb=False)
 
         # learning stats
-        if "policy_loss" in results:
+        if 'policy_loss' in results:
             self.logger.add_scalars(
                 {
                     k: results[k]
-                    for k in ["policy_loss", "critic_loss"]
+                    for k in ['policy_loss', 'critic_loss']
                 },
                 step,
-                prefix="loss")
+                prefix='loss')
 
         # performance stats
         ep_lengths = np.asarray(self.env.length_queue)
         ep_returns = np.asarray(self.env.return_queue)
-        ep_constraint_violation = np.asarray(self.env.queued_stats["constraint_violation"])
+        ep_constraint_violation = np.asarray(self.env.queued_stats['constraint_violation'])
         self.logger.add_scalars(
             {
-                "ep_length": ep_lengths.mean(),
-                "ep_return": ep_returns.mean(),
-                "ep_reward": (ep_returns / ep_lengths).mean(),
-                "ep_constraint_violation": ep_constraint_violation.mean()
+                'ep_length': ep_lengths.mean(),
+                'ep_return': ep_returns.mean(),
+                'ep_reward': (ep_returns / ep_lengths).mean(),
+                'ep_constraint_violation': ep_constraint_violation.mean()
             },
             step,
-            prefix="stat")
+            prefix='stat')
 
         # total constraint violation during learning
-        total_violations = self.env.accumulated_stats["constraint_violation"]
-        self.logger.add_scalars({"constraint_violation": total_violations}, step, prefix="stat")
+        total_violations = self.env.accumulated_stats['constraint_violation']
+        self.logger.add_scalars({'constraint_violation': total_violations}, step, prefix='stat')
 
-        if "eval" in results:
-            eval_ep_lengths = results["eval"]["ep_lengths"]
-            eval_ep_returns = results["eval"]["ep_returns"]
-            eval_constraint_violation = results["eval"]["constraint_violation"]
-            eval_mse = results["eval"]["mse"]
+        if 'eval' in results:
+            eval_ep_lengths = results['eval']['ep_lengths']
+            eval_ep_returns = results['eval']['ep_returns']
+            eval_constraint_violation = results['eval']['constraint_violation']
+            eval_mse = results['eval']['mse']
             self.logger.add_scalars(
                 {
-                    "ep_length": eval_ep_lengths.mean(),
-                    "ep_return": eval_ep_returns.mean(),
-                    "ep_reward": (eval_ep_returns / eval_ep_lengths).mean(),
-                    "constraint_violation": eval_constraint_violation.mean(),
-                    "mse": eval_mse.mean()
+                    'ep_length': eval_ep_lengths.mean(),
+                    'ep_return': eval_ep_returns.mean(),
+                    'ep_reward': (eval_ep_returns / eval_ep_lengths).mean(),
+                    'constraint_violation': eval_constraint_violation.mean(),
+                    'mse': eval_mse.mean()
                 },
                 step,
-                prefix="stat_eval")
+                prefix='stat_eval')
 
         # print summary table
         self.logger.dump_scalars()
