@@ -1,25 +1,26 @@
-"""Utility functions for Gaussian Processes.
+'''Utility functions for Gaussian Processes.'''
 
-"""
 import os.path
+from copy import deepcopy
+
 import numpy as np
 import gpytorch
 import torch
 import matplotlib.pyplot as plt
 import casadi as ca
-from copy import deepcopy
 from sklearn import preprocessing
 from sklearn.cluster import KMeans
 
 from safe_control_gym.utils.utils import mkdirs
 torch.manual_seed(0)
 
+
 def covSEard(x,
              z,
              ell,
              sf2
              ):
-    """GP squared exponential kernel.
+    '''GP squared exponential kernel.
 
     This function is based on the 2018 GP-MPC library by Helge-André Langåker
 
@@ -31,18 +32,16 @@ def covSEard(x,
 
     Returns:
         SE kernel (casadi.MX/SX): SE kernel.
-
-    """
+    '''
     dist = ca.sum1((x - z)**2 / ell**2)
     return sf2 * ca.SX.exp(-.5 * dist)
 
 
 class ZeroMeanIndependentMultitaskGPModel(gpytorch.models.ExactGP):
-    """Multidimensional Gaussian Process model with zero mean function.
+    '''Multidimensional Gaussian Process model with zero mean function.
 
     Or constant mean and radial basis function kernel (SE).
-
-    """
+    '''
 
     def __init__(self,
                  train_x,
@@ -50,15 +49,14 @@ class ZeroMeanIndependentMultitaskGPModel(gpytorch.models.ExactGP):
                  likelihood,
                  nx
                  ):
-        """Initialize a multidimensional Gaussian Process model with zero mean function.
+        '''Initialize a multidimensional Gaussian Process model with zero mean function.
 
         Args:
             train_x (torch.Tensor): input training data (input_dim X N samples).
             train_y (torch.Tensor): output training data (output dim x N samples).
             likelihood (gpytorch.likelihood): Likelihood function (gpytorch.likelihoods.MultitaskGaussianLikelihood).
             nx (int): dimension of the target output (output dim)
-
-        """
+        '''
         super().__init__(train_x, train_y, likelihood)
         self.n = nx
         # For Zero mean function.
@@ -75,9 +73,6 @@ class ZeroMeanIndependentMultitaskGPModel(gpytorch.models.ExactGP):
     def forward(self,
                 x
                 ):
-        """
-
-        """
         mean_x = self.mean_module(x)
         covar_x = self.covar_module(x)
         return gpytorch.distributions.MultitaskMultivariateNormal.from_batch_mvn(
@@ -86,24 +81,23 @@ class ZeroMeanIndependentMultitaskGPModel(gpytorch.models.ExactGP):
 
 
 class ZeroMeanIndependentGPModel(gpytorch.models.ExactGP):
-    """Single dimensional output Gaussian Process model with zero mean function.
+    '''Single dimensional output Gaussian Process model with zero mean function.
 
     Or constant mean and radial basis function kernel (SE).
+    '''
 
-    """
     def __init__(self,
                  train_x,
                  train_y,
                  likelihood
                  ):
-        """Initialize a single dimensional Gaussian Process model with zero mean function.
+        '''Initialize a single dimensional Gaussian Process model with zero mean function.
 
         Args:
             train_x (torch.Tensor): input training data (input_dim X N samples).
             train_y (torch.Tensor): output training data (output dim x N samples).
             likelihood (gpytorch.likelihood): Likelihood function (gpytorch.likelihoods.GaussianLikelihood).
-
-        """
+        '''
         super().__init__(train_x, train_y, likelihood)
         # For Zero mean function.
         self.mean_module = gpytorch.means.ZeroMean()
@@ -116,18 +110,13 @@ class ZeroMeanIndependentGPModel(gpytorch.models.ExactGP):
     def forward(self,
                 x
                 ):
-        """
-
-        """
         mean_x = self.mean_module(x)
         covar_x = self.covar_module(x)
         return gpytorch.distributions.MultivariateNormal(mean_x, covar_x)
 
 
 class GaussianProcessCollection:
-    """Collection of GaussianProcesses for multioutput GPs.
-
-    """
+    '''Collection of GaussianProcesses for multioutput GPs.'''
 
     def __init__(self, model_type,
                  likelihood,
@@ -136,7 +125,7 @@ class GaussianProcessCollection:
                  target_mask=None,
                  normalize=False
                  ):
-        """Creates a single GaussianProcess for each output dimension.
+        '''Creates a single GaussianProcess for each output dimension.
 
         Args:
             model_type (gpytorch model class): Model class for the GP (ZeroMeanIndependentGPModel).
@@ -145,8 +134,7 @@ class GaussianProcessCollection:
             input_mask (list): Input dimensions to keep. If None, use all input dimensions.
             target_mask (list): Target dimensions to keep. If None, use all target dimensions.
             normalize (bool): If True, scale all data between -1 and 1.
-
-        """
+        '''
         self.gp_list = []
         self.model_type = model_type
         self.likelihood = likelihood
@@ -155,7 +143,7 @@ class GaussianProcessCollection:
         self.NORMALIZE = normalize
         self.input_mask = input_mask
         self.target_mask = target_mask
-        for i in range(target_dim):
+        for _ in range(target_dim):
             self.gp_list.append(GaussianProcess(model_type,
                                                 deepcopy(likelihood),
                                                 input_mask=input_mask,
@@ -165,12 +153,11 @@ class GaussianProcessCollection:
                          train_inputs,
                          train_targets
                          ):
-        """Initialize useful properties.
+        '''Initialize useful properties.
 
         Args:
             train_inputs, train_targets (torch.tensors): Input and target training data.
-
-        """
+        '''
         target_dimension = train_targets.shape[1]
         self.input_dimension = train_inputs.shape[1]
         self.output_dimension = target_dimension
@@ -181,25 +168,23 @@ class GaussianProcessCollection:
                              train_targets,
                              path_to_statedicts
                              ):
-        """Load hyperparameters from a state_dict.
+        '''Load hyperparameters from a state_dict.
 
         Args:
             train_inputs, train_targets (torch.tensors): Input and target training data.
             path_to_statedicts (str): Path to where the state dicts are saved.
-
-        """
+        '''
         self._init_properties(train_inputs, train_targets)
-        target_dimension = train_targets.shape[1]
         gp_K_plus_noise_list = []
         gp_K_plus_noise_inv_list = []
         for gp_ind, gp in enumerate(self.gp_list):
-            path = os.path.join(path_to_statedicts, 'best_model_%s.pth'  % self.target_mask[gp_ind])
-            print("#########################################")
-            print("#       Loading GP dimension %s         #" % self.target_mask[gp_ind])
-            print("#########################################")
-            print('Path: %s' % path)
+            path = os.path.join(path_to_statedicts, f'best_model_{self.target_mask[gp_ind]}.pth')
+            print('#########################################')
+            print('#       Loading GP dimension {self.target_mask[gp_ind]}         #')
+            print('#########################################')
+            print(f'Path: {path}')
             gp.init_with_hyperparam(train_inputs,
-                                    train_targets[:,self.target_mask[gp_ind]],
+                                    train_targets[:, self.target_mask[gp_ind]],
                                     path)
             gp_K_plus_noise_list.append(gp.model.K_plus_noise.detach())
             gp_K_plus_noise_inv_list.append(gp.model.K_plus_noise_inv.detach())
@@ -213,9 +198,7 @@ class GaussianProcessCollection:
     def get_hyperparameters(self,
                             as_numpy=False
                             ):
-        """Get the outputscale and lengthscale from the kernel matrices of the GPs.
-
-        """
+        '''Get the outputscale and lengthscale from the kernel matrices of the GPs.'''
         lengthscale_list = []
         output_scale_list = []
         noise_list = []
@@ -239,37 +222,36 @@ class GaussianProcessCollection:
               n_train=[500],
               learning_rate=[0.01],
               gpu=False,
-              dir='results'
+              output_dir='results'
               ):
-        """Train the GP using Train_x and Train_y.
+        '''Train the GP using Train_x and Train_y.
 
         Args:
             train_x: Torch tensor (N samples [rows] by input dim [cols])
             train_y: Torch tensor (N samples [rows] by target dim [cols])
-
-        """
+        '''
         self._init_properties(train_x_raw, train_y_raw)
         self.model_paths = []
-        mkdirs(dir)
+        mkdirs(output_dir)
         gp_K_plus_noise_inv_list = []
         gp_K_plus_noise_list = []
         for gp_ind, gp in enumerate(self.gp_list):
             lr = learning_rate[self.target_mask[gp_ind]]
             n_t = n_train[self.target_mask[gp_ind]]
-            print("#########################################")
-            print("#      Training GP dimension %s         #" % self.target_mask[gp_ind])
-            print("#########################################")
-            print("Train iterations: %s" % n_t)
-            print("Learning Rate: %s" % lr)
+            print('#########################################')
+            print(f'#      Training GP dimension {self.target_mask[gp_ind]}         #')
+            print('#########################################')
+            print(f'Train iterations: {n_t}')
+            print(f'Learning Rate: {lr}')
             gp.train(train_x_raw,
-                     train_y_raw[:,self.target_mask[gp_ind]],
+                     train_y_raw[:, self.target_mask[gp_ind]],
                      test_x_raw,
                      test_y_raw[:, self.target_mask[gp_ind]],
                      n_train=n_t,
                      learning_rate=lr,
                      gpu=gpu,
-                     fname=os.path.join(dir, 'best_model_%s.pth' % self.target_mask[gp_ind]))
-            self.model_paths.append(dir)
+                     fname=os.path.join(output_dir, f'best_model_{self.target_mask[gp_ind]}.pth'))
+            self.model_paths.append(output_dir)
             gp_K_plus_noise_list.append(gp.model.K_plus_noise)
             gp_K_plus_noise_inv_list.append(gp.model.K_plus_noise_inv)
         gp_K_plus_noise = torch.stack(gp_K_plus_noise_list)
@@ -278,14 +260,12 @@ class GaussianProcessCollection:
         self.K_plus_noise_inv = gp_K_plus_noise_inv
         self.casadi_predict = self.make_casadi_predict_func()
 
-
     def predict(self,
                 x,
                 requires_grad=False,
                 return_pred=True
                 ):
-        """
-
+        '''
         Args:
             x : torch.Tensor (N_samples x input DIM).
 
@@ -294,8 +274,7 @@ class GaussianProcessCollection:
                 mean : torch.tensor (nx X N_samples).
                 lower : torch.tensor (nx X N_samples).
                 upper : torch.tensor (nx X N_samples).
-
-        """
+        '''
         means_list = []
         cov_list = []
         pred_list = []
@@ -315,11 +294,10 @@ class GaussianProcessCollection:
             return means, cov
 
     def make_casadi_predict_func(self):
-        """
+        '''
         Assume train_inputs and train_tergets are already
-        """
+        '''
 
-        means_list = []
         Nz = len(self.input_mask)
         Ny = len(self.target_mask)
         z = ca.SX.sym('z1', Nz)
@@ -333,14 +311,10 @@ class GaussianProcessCollection:
                                      ['mean'])
         return casadi_predict
 
-
-
     def prediction_jacobian(self,
                             query
                             ):
-        """Return Jacobian.
-
-        """
+        '''Return Jacobian.'''
         raise NotImplementedError
 
     def plot_trained_gp(self,
@@ -348,12 +322,10 @@ class GaussianProcessCollection:
                         targets,
                         fig_count=0
                         ):
-        """Plot the trained GP given the input and target data.
-
-        """
+        '''Plot the trained GP given the input and target data.'''
         for gp_ind, gp in enumerate(self.gp_list):
             fig_count = gp.plot_trained_gp(inputs,
-                                           targets[:,self.target_mask[gp_ind],None],
+                                           targets[:, self.target_mask[gp_ind], None],
                                            self.target_mask[gp_ind],
                                            fig_count=fig_count)
             fig_count += 1
@@ -362,7 +334,7 @@ class GaussianProcessCollection:
                      x1,
                      x2=None
                      ):
-        """Evaluate the kernel given vectors x1 and x2.
+        '''Evaluate the kernel given vectors x1 and x2.
 
         Args:
             x1 (torch.Tensor): First vector.
@@ -370,12 +342,11 @@ class GaussianProcessCollection:
 
         Returns:
             list of LazyTensor Kernels.
-
-        """
+        '''
         if x2 is None:
             x2 = x1
-        # todo: Make normalization at the GPCollection level?
-        #if self.NORMALIZE:
+        # TODO: Make normalization at the GPCollection level?
+        # if self.NORMALIZE:
         #    x1 = torch.from_numpy(self.gp_list[0].scaler.transform(x1.numpy()))
         #    x2 = torch.from_numpy(self.gp_list[0].scaler.transform(x2.numpy()))
         k_list = []
@@ -387,7 +358,7 @@ class GaussianProcessCollection:
                x1,
                x2=None
                ):
-        """Evaluate the kernel given vectors x1 and x2.
+        '''Evaluate the kernel given vectors x1 and x2.
 
         Args:
             x1 (torch.Tensor): First vector.
@@ -395,8 +366,7 @@ class GaussianProcessCollection:
 
         Returns:
             Torch tensor of the non-lazy kernel matrices.
-
-        """
+        '''
         k_list = self._kernel_list(x1, x2)
         non_lazy_tensors = [k.evaluate() for k in k_list]
         return torch.stack(non_lazy_tensors)
@@ -405,7 +375,7 @@ class GaussianProcessCollection:
                    x1,
                    x2=None
                    ):
-        """Evaluate the inverse kernel given vectors x1 and x2.
+        '''Evaluate the inverse kernel given vectors x1 and x2.
 
         Only works for square kernel.
 
@@ -415,37 +385,34 @@ class GaussianProcessCollection:
 
         Returns:
             Torch tensor of the non-lazy inverse kernel matrices.
-
-        """
+        '''
         if x2 is None:
             x2 = x1
-        assert x1.shape == x2.shape, ValueError("x1 and x2 need to have the same shape.")
+        assert x1.shape == x2.shape, ValueError('x1 and x2 need to have the same shape.')
         k_list = self._kernel_list(x1, x2)
         num_of_points = x1.shape[0]
         # Efficient inversion is performed VIA inv_matmul on the laze tensor with Identity.
         non_lazy_tensors = [k.inv_matmul(torch.eye(num_of_points).double()) for k in k_list]
         return torch.stack(non_lazy_tensors)
 
-class GaussianProcess:
-    """Gaussian Process decorator for gpytorch.
 
-    """
+class GaussianProcess:
+    '''Gaussian Process decorator for gpytorch.'''
 
     def __init__(self,
-                  model_type,
-                  likelihood,
-                  input_mask=None,
-                  target_mask=None,
-                  normalize=False
-                  ):
-        """Initialize Gaussian Process.
+                 model_type,
+                 likelihood,
+                 input_mask=None,
+                 target_mask=None,
+                 normalize=False
+                 ):
+        '''Initialize Gaussian Process.
 
         Args:
             model_type (gpytorch model class): Model class for the GP (ZeroMeanIndependentMultitaskGPModel).
             likelihood (gpytorch.likelihood): likelihood function.
             normalize (bool): If True, scale all data between -1 and 1. (prototype and not fully operational).
-
-        """
+        '''
         self.model_type = model_type
         self.likelihood = likelihood
         self.optimizer = None
@@ -458,9 +425,7 @@ class GaussianProcess:
                     train_inputs,
                     train_targets
                     ):
-        """Init GP model from train inputs and train_targets.
-
-        """
+        '''Init GP model from train inputs and train_targets.'''
         if train_targets.ndim > 1:
             target_dimension = train_targets.shape[1]
         else:
@@ -480,29 +445,24 @@ class GaussianProcess:
         self.output_dimension = target_dimension
         self.n_training_samples = train_inputs.shape[0]
 
-
     def _compute_GP_covariances(self,
                                 train_x
                                 ):
-        """Compute K(X,X) + sigma*I and its inverse.
-
-        """
+        '''Compute K(X,X) + sigma*I and its inverse.'''
         # Pre-compute inverse covariance plus noise to speed-up computation.
         K_lazy = self.model.covar_module(train_x.double())
         K_lazy_plus_noise = K_lazy.add_diag(self.model.likelihood.noise)
         n_samples = train_x.shape[0]
         self.model.K_plus_noise = K_lazy_plus_noise.matmul(torch.eye(n_samples).double())
         self.model.K_plus_noise_inv = K_lazy_plus_noise.inv_matmul(torch.eye(n_samples).double())
-        #self.model.K_plus_noise_inv_2 = torch.inverse(self.model.K_plus_noise) # Equivalent to above but slower.
+        # self.model.K_plus_noise_inv_2 = torch.inverse(self.model.K_plus_noise) # Equivalent to above but slower.
 
     def init_with_hyperparam(self,
                              train_inputs,
                              train_targets,
                              path_to_statedict
                              ):
-        """Load hyperparameters from a state_dict.
-
-        """
+        '''Load hyperparameters from a state_dict.'''
         if self.input_mask is not None:
             train_inputs = train_inputs[:, self.input_mask]
         if self.target_mask is not None:
@@ -513,7 +473,7 @@ class GaussianProcess:
         if self.NORMALIZE:
             train_inputs = torch.from_numpy(self.scaler.transform(train_inputs.numpy()))
         self.model.load_state_dict(state_dict)
-        self.model.double() # needed otherwise loads state_dict as float32
+        self.model.double()  # needed otherwise loads state_dict as float32
         self._compute_GP_covariances(train_inputs)
         self.casadi_predict = self.make_casadi_prediction_func(train_inputs, train_targets)
 
@@ -527,13 +487,12 @@ class GaussianProcess:
               gpu=False,
               fname='best_model.pth',
               ):
-        """Train the GP using Train_x and Train_y.
+        '''Train the GP using Train_x and Train_y.
 
         Args:
             train_x: Torch tensor (N samples [rows] by input dim [cols])
             train_y: Torch tensor (N samples [rows] by target dim [cols])
-
-        """
+        '''
         train_x_raw = train_input_data
         train_y_raw = train_target_data
         test_x_raw = test_input_data
@@ -588,7 +547,7 @@ class GaussianProcess:
                 print('Iter %d/%d - MLL trian Loss: %.3f, Posterior Test Loss: %0.3f' % (i + 1, n_train, loss.item(), test_loss.item()))
 
             self.optimizer.step()
-            #if test_loss < best_loss:
+            # if test_loss < best_loss:
             #    best_loss = test_loss
             #    state_dict = self.model.state_dict()
             #    torch.save(state_dict, fname)
@@ -599,10 +558,10 @@ class GaussianProcess:
                 torch.save(state_dict, fname)
                 best_epoch = i
 
-            i+=1
-        print("Training Complete")
-        print("Lowest epoch: %s" %best_epoch)
-        print("Lowest Loss: %s" % best_loss)
+            i += 1
+        print('Training Complete')
+        print(f'Lowest epoch: {best_epoch}')
+        print(f'Lowest Loss: {best_loss}')
         self.model = self.model.cpu()
         self.likelihood = self.likelihood.cpu()
         train_x = train_x.cpu()
@@ -616,8 +575,7 @@ class GaussianProcess:
                 requires_grad=False,
                 return_pred=True
                 ):
-        """
-
+        '''
         Args:
             x : torch.Tensor (N_samples x input DIM).
 
@@ -626,14 +584,13 @@ class GaussianProcess:
                 mean : torch.tensor (nx X N_samples).
                 lower : torch.tensor (nx X N_samples).
                 upper : torch.tensor (nx X N_samples).
-
-        """
+        '''
         self.model.eval()
         self.likelihood.eval()
-        if type(x) is np.ndarray:
+        if isinstance(x, np.ndarray):
             x = torch.from_numpy(x).double()
         if self.input_mask is not None:
-            x = x[:,self.input_mask]
+            x = x[:, self.input_mask]
         if self.NORMALIZE:
             x = torch.from_numpy(self.scaler.transform(x))
         if requires_grad:
@@ -650,19 +607,16 @@ class GaussianProcess:
         else:
             return mean, cov
 
-
     def prediction_jacobian(self,
                             query
                             ):
-        mean_der, cov_der = torch.autograd.functional.jacobian(
-                                lambda x: self.predict(x, requires_grad=True, return_pred=False),
-                                query.double())
+        mean_der, _ = torch.autograd.functional.jacobian(
+            lambda x: self.predict(x, requires_grad=True, return_pred=False),
+            query.double())
         return mean_der.detach().squeeze()
 
     def make_casadi_prediction_func(self, train_inputs, train_targets):
-        """
-        Assumes train_inputs and train_targets are already masked.
-        """
+        '''Assumes train_inputs and train_targets are already masked.'''
         train_inputs = train_inputs.numpy()
         train_targets = train_targets.numpy()
         lengthscale = self.model.covar_module.base_kernel.lengthscale.detach().numpy()
@@ -689,14 +643,14 @@ class GaussianProcess:
                         ):
         if self.target_mask is not None:
             targets = targets[:, self.target_mask]
-        means, covs, preds = self.predict(inputs)
+        means, _, preds = self.predict(inputs)
         t = np.arange(inputs.shape[0])
         lower, upper = preds.confidence_region()
         for i in range(self.output_dimension):
             fig_count += 1
             plt.figure(fig_count)
             if lower.ndim > 1:
-                plt.fill_between(t, lower[:,i].detach().numpy(), upper[:,i].detach().numpy(), alpha=0.5, label='95%')
+                plt.fill_between(t, lower[:, i].detach().numpy(), upper[:, i].detach().numpy(), alpha=0.5, label='95%')
                 plt.plot(t, means[:, i], 'r', label='GP Mean')
                 plt.plot(t, targets[:, i], '*k', label='Data')
             else:
@@ -704,14 +658,15 @@ class GaussianProcess:
                 plt.plot(t, means, 'r', label='GP Mean')
                 plt.plot(t, targets, '*k', label='Data')
             plt.legend()
-            plt.title('Fitted GP x%s' % output_label)
+            plt.title(f'Fitted GP x{output_label}')
             plt.xlabel('Time (s)')
             plt.ylabel('v')
             plt.show()
         return fig_count
 
+
 def kmeans_centriods(n_cent, data, rand_state=0):
-    """kmeans clustering. Useful for finding reasonable inducing points.
+    '''kmeans clustering. Useful for finding reasonable inducing points.
 
     Args:
         n_cent (int): Number of centriods.
@@ -719,7 +674,6 @@ def kmeans_centriods(n_cent, data, rand_state=0):
 
     Return:
         centriods (np.array): Array of centriods (n_cent X n_features).
-
-    """
+    '''
     kmeans = KMeans(n_clusters=n_cent, random_state=rand_state).fit(data)
     return kmeans.cluster_centers_
