@@ -1,4 +1,4 @@
-"""Proximal Policy Optimization (PPO)
+'''Proximal Policy Optimization (PPO)
 
 Based on:
     * https://github.com/openai/spinningup/blob/master/spinup/algos/pytorch/ppo/ppo.py
@@ -10,7 +10,7 @@ Additional references:
     * pytorch-a2c-ppo-acktr-gail - https://github.com/ikostrikov/pytorch-a2c-ppo-acktr-gail
     * openai spinning up - ppo - https://github.com/openai/spinningup/tree/master/spinup/algos/pytorch/ppo
     * stable baselines3 - ppo - https://github.com/DLR-RM/stable-baselines3/tree/master/stable_baselines3/ppo
-"""
+'''
 
 import os
 import time
@@ -28,15 +28,13 @@ from safe_control_gym.controllers.ppo.ppo_utils import PPOAgent, PPOBuffer, comp
 
 
 class PPO(BaseController):
-    """Proximal policy optimization.
-
-    """
+    '''Proximal policy optimization.'''
 
     def __init__(self,
                  env_func,
                  training=True,
-                 checkpoint_path="model_latest.pt",
-                 output_dir="temp",
+                 checkpoint_path='model_latest.pt',
+                 output_dir='temp',
                  use_gpu=False,
                  seed=0,
                  **kwargs):
@@ -83,29 +81,25 @@ class PPO(BaseController):
         self.logger = ExperimentLogger(output_dir, log_file_out=log_file_out, use_tensorboard=use_tensorboard)
 
     def reset(self):
-        """Do initializations for training or evaluation.
-
-        """
+        '''Do initializations for training or evaluation.'''
         if self.training:
             # set up stats tracking
-            self.env.add_tracker("constraint_violation", 0)
-            self.env.add_tracker("constraint_violation", 0, mode="queue")
-            self.eval_env.add_tracker("constraint_violation", 0, mode="queue")
-            self.eval_env.add_tracker("mse", 0, mode="queue")
+            self.env.add_tracker('constraint_violation', 0)
+            self.env.add_tracker('constraint_violation', 0, mode='queue')
+            self.eval_env.add_tracker('constraint_violation', 0, mode='queue')
+            self.eval_env.add_tracker('mse', 0, mode='queue')
 
             self.total_steps = 0
             obs, _ = self.env.reset()
             self.obs = self.obs_normalizer(obs)
         else:
             # Add episodic stats to be tracked.
-            self.env.add_tracker("constraint_violation", 0, mode="queue")
-            self.env.add_tracker("constraint_values", 0, mode="queue")
-            self.env.add_tracker("mse", 0, mode="queue")
+            self.env.add_tracker('constraint_violation', 0, mode='queue')
+            self.env.add_tracker('constraint_values', 0, mode='queue')
+            self.env.add_tracker('mse', 0, mode='queue')
 
     def close(self):
-        """Shuts down and cleans up lingering resources.
-
-        """
+        '''Shuts down and cleans up lingering resources.'''
         self.env.close()
         if self.training:
             self.eval_env.close()
@@ -114,22 +108,20 @@ class PPO(BaseController):
     def save(self,
              path
              ):
-        """Saves model params and experiment state to checkpoint path.
-
-        """
+        '''Saves model params and experiment state to checkpoint path.'''
         path_dir = os.path.dirname(path)
         os.makedirs(path_dir, exist_ok=True)
         state_dict = {
-            "agent": self.agent.state_dict(),
-            "obs_normalizer": self.obs_normalizer.state_dict(),
-            "reward_normalizer": self.reward_normalizer.state_dict(),
+            'agent': self.agent.state_dict(),
+            'obs_normalizer': self.obs_normalizer.state_dict(),
+            'reward_normalizer': self.reward_normalizer.state_dict(),
         }
         if self.training:
             exp_state = {
-                "total_steps": self.total_steps,
-                "obs": self.obs,
-                "random_state": get_random_state(),
-                "env_random_state": self.env.get_env_random_state()
+                'total_steps': self.total_steps,
+                'obs': self.obs,
+                'random_state': get_random_state(),
+                'env_random_state': self.env.get_env_random_state()
             }
             state_dict.update(exp_state)
         torch.save(state_dict, path)
@@ -137,60 +129,56 @@ class PPO(BaseController):
     def load(self,
              path
              ):
-        """Restores model and experiment given checkpoint path.
-
-        """
+        '''Restores model and experiment given checkpoint path.'''
         state = torch.load(path)
         # Restore policy.
-        self.agent.load_state_dict(state["agent"])
-        self.obs_normalizer.load_state_dict(state["obs_normalizer"])
-        self.reward_normalizer.load_state_dict(state["reward_normalizer"])
+        self.agent.load_state_dict(state['agent'])
+        self.obs_normalizer.load_state_dict(state['obs_normalizer'])
+        self.reward_normalizer.load_state_dict(state['reward_normalizer'])
         # Restore experiment state.
         if self.training:
-            self.total_steps = state["total_steps"]
-            self.obs = state["obs"]
-            set_random_state(state["random_state"])
-            self.env.set_env_random_state(state["env_random_state"])
+            self.total_steps = state['total_steps']
+            self.obs = state['obs']
+            set_random_state(state['random_state'])
+            self.env.set_env_random_state(state['env_random_state'])
             self.logger.load(self.total_steps)
 
     def learn(self,
               env=None,
               **kwargs
               ):
-        """Performs learning (pre-training, training, fine-tuning, etc).
-
-        """
+        '''Performs learning (pre-training, training, fine-tuning, etc).'''
         while self.total_steps < self.max_env_steps:
             results = self.train_step()
             # Checkpoint.
             if self.total_steps >= self.max_env_steps or (self.save_interval and self.total_steps % self.save_interval == 0):
                 # Latest/final checkpoint.
                 self.save(self.checkpoint_path)
-                self.logger.info("Checkpoint | {}".format(self.checkpoint_path))
+                self.logger.info(f'Checkpoint | {self.checkpoint_path}')
             if self.num_checkpoints and self.total_steps % (self.max_env_steps // self.num_checkpoints) == 0:
                 # Intermediate checkpoint.
-                path = os.path.join(self.output_dir, "checkpoints", "model_{}.pt".format(self.total_steps))
+                path = os.path.join(self.output_dir, 'checkpoints', f'model_{self.total_steps}.pt')
                 self.save(path)
             # Evaluation.
             if self.eval_interval and self.total_steps % self.eval_interval == 0:
                 eval_results = self.run(env=self.eval_env, n_episodes=self.eval_batch_size)
-                results["eval"] = eval_results
-                self.logger.info("Eval | ep_lengths {:.2f} +/- {:.2f} | ep_return {:.3f} +/- {:.3f}".format(eval_results["ep_lengths"].mean(),
-                                                                                                            eval_results["ep_lengths"].std(),
-                                                                                                            eval_results["ep_returns"].mean(),
-                                                                                                            eval_results["ep_returns"].std()))
+                results['eval'] = eval_results
+                self.logger.info('Eval | ep_lengths {:.2f} +/- {:.2f} | ep_return {:.3f} +/- {:.3f}'.format(eval_results['ep_lengths'].mean(),
+                                                                                                            eval_results['ep_lengths'].std(),
+                                                                                                            eval_results['ep_returns'].mean(),
+                                                                                                            eval_results['ep_returns'].std()))
                 # Save best model.
-                eval_score = eval_results["ep_returns"].mean()
-                eval_best_score = getattr(self, "eval_best_score", -np.infty)
+                eval_score = eval_results['ep_returns'].mean()
+                eval_best_score = getattr(self, 'eval_best_score', -np.infty)
                 if self.eval_save_best and eval_best_score < eval_score:
                     self.eval_best_score = eval_score
-                    self.save(os.path.join(self.output_dir, "model_best.pt"))
+                    self.save(os.path.join(self.output_dir, 'model_best.pt'))
             # Logging.
             if self.log_interval and self.total_steps % self.log_interval == 0:
                 self.log_step(results)
 
     def select_action(self, obs, info=None):
-        """Determine the action to take at the current timestep.
+        '''Determine the action to take at the current timestep.
 
         Args:
             obs (ndarray): The observation at this timestep.
@@ -198,7 +186,7 @@ class PPO(BaseController):
 
         Returns:
             action (ndarray): The action chosen by the controller.
-        """
+        '''
 
         with torch.no_grad():
             obs = torch.FloatTensor(obs).to(self.device)
@@ -211,11 +199,8 @@ class PPO(BaseController):
             render=False,
             n_episodes=10,
             verbose=False,
-            **kwargs
             ):
-        """Runs evaluation with current policy.
-
-        """
+        '''Runs evaluation with current policy.'''
         self.agent.eval()
         self.obs_normalizer.set_read_only()
         if env is None:
@@ -224,9 +209,9 @@ class PPO(BaseController):
             if not is_wrapped(env, RecordEpisodeStatistics):
                 env = RecordEpisodeStatistics(env, n_episodes)
                 # Add episodic stats to be tracked.
-                env.add_tracker("constraint_violation", 0, mode="queue")
-                env.add_tracker("constraint_values", 0, mode="queue")
-                env.add_tracker("mse", 0, mode="queue")
+                env.add_tracker('constraint_violation', 0, mode='queue')
+                env.add_tracker('constraint_values', 0, mode='queue')
+                env.add_tracker('mse', 0, mode='queue')
 
         obs, info = env.reset()
         obs = self.obs_normalizer(obs)
@@ -234,24 +219,24 @@ class PPO(BaseController):
         frames = []
         while len(ep_returns) < n_episodes:
             action = self.select_action(obs=obs, info=info)
-            obs, reward, done, info = env.step(action)
+            obs, _, done, info = env.step(action)
             if render:
                 env.render()
-                frames.append(env.render("rgb_array"))
+                frames.append(env.render('rgb_array'))
             if verbose:
-                print("obs {} | act {}".format(obs, action))
+                print(f'obs {obs} | act {action}')
             if done:
-                assert "episode" in info
-                ep_returns.append(info["episode"]["r"])
-                ep_lengths.append(info["episode"]["l"])
+                assert 'episode' in info
+                ep_returns.append(info['episode']['r'])
+                ep_lengths.append(info['episode']['l'])
                 obs, _ = env.reset()
             obs = self.obs_normalizer(obs)
         # Collect evaluation results.
         ep_lengths = np.asarray(ep_lengths)
         ep_returns = np.asarray(ep_returns)
-        eval_results = {"ep_returns": ep_returns, "ep_lengths": ep_lengths}
+        eval_results = {'ep_returns': ep_returns, 'ep_lengths': ep_lengths}
         if len(frames) > 0:
-            eval_results["frames"] = frames
+            eval_results['frames'] = frames
         # Other episodic stats from evaluation env.
         if len(env.queued_stats) > 0:
             queued_stats = {k: np.asarray(v) for k, v in env.queued_stats.items()}
@@ -259,15 +244,13 @@ class PPO(BaseController):
         return eval_results
 
     def train_step(self):
-        """Performs a training/fine-tuning step.
-
-        """
+        '''Performs a training/fine-tuning step.'''
         self.agent.train()
         self.obs_normalizer.unset_read_only()
         rollouts = PPOBuffer(self.env.observation_space, self.env.action_space, self.rollout_steps, self.rollout_batch_size)
         obs = self.obs
         start = time.time()
-        for step in range(self.rollout_steps):
+        for _ in range(self.rollout_steps):
             with torch.no_grad():
                 act, v, logp = self.agent.ac.step(torch.FloatTensor(obs).to(self.device))
             next_obs, rew, done, info = self.env.step(act)
@@ -276,16 +259,16 @@ class PPO(BaseController):
             mask = 1 - done.astype(float)
             # Time truncation is not the same as true termination.
             terminal_v = np.zeros_like(v)
-            for idx, inf in enumerate(info["n"]):
-                if "terminal_info" not in inf:
+            for idx, inf in enumerate(info['n']):
+                if 'terminal_info' not in inf:
                     continue
-                inff = inf["terminal_info"]
-                if "TimeLimit.truncated" in inff and inff["TimeLimit.truncated"]:
-                    terminal_obs = inf["terminal_observation"]
+                inff = inf['terminal_info']
+                if 'TimeLimit.truncated' in inff and inff['TimeLimit.truncated']:
+                    terminal_obs = inf['terminal_observation']
                     terminal_obs_tensor = torch.FloatTensor(terminal_obs).unsqueeze(0).to(self.device)
                     terminal_val = self.agent.ac.critic(terminal_obs_tensor).squeeze().detach().cpu().numpy()
                     terminal_v[idx] = terminal_val
-            rollouts.push({"obs": obs, "act": act, "rew": rew, "mask": mask, "v": v, "logp": logp, "terminal_v": terminal_v})
+            rollouts.push({'obs': obs, 'act': act, 'rew': rew, 'mask': mask, 'v': v, 'logp': logp, 'terminal_v': terminal_v})
             obs = next_obs
         self.obs = obs
         self.total_steps += self.rollout_batch_size * self.rollout_steps
@@ -303,65 +286,63 @@ class PPO(BaseController):
         # Prevent divide-by-0 for repetitive tasks.
         rollouts.adv = (adv - adv.mean()) / (adv.std() + 1e-6)
         results = self.agent.update(rollouts, self.device)
-        results.update({"step": self.total_steps, "elapsed_time": time.time() - start})
+        results.update({'step': self.total_steps, 'elapsed_time': time.time() - start})
         return results
 
     def log_step(self,
                  results
                  ):
-        """Does logging after a training step.
-
-        """
-        step = results["step"]
+        '''Does logging after a training step.'''
+        step = results['step']
         # runner stats
         self.logger.add_scalars(
             {
-                "step": step,
-                "step_time": results["elapsed_time"],
-                "progress": step / self.max_env_steps
+                'step': step,
+                'step_time': results['elapsed_time'],
+                'progress': step / self.max_env_steps
             },
             step,
-            prefix="time",
+            prefix='time',
             write=False,
             write_tb=False)
         # Learning stats.
         self.logger.add_scalars(
             {
                 k: results[k]
-                for k in ["policy_loss", "value_loss", "entropy_loss", "approx_kl"]
+                for k in ['policy_loss', 'value_loss', 'entropy_loss', 'approx_kl']
             },
             step,
-            prefix="loss")
+            prefix='loss')
         # Performance stats.
         ep_lengths = np.asarray(self.env.length_queue)
         ep_returns = np.asarray(self.env.return_queue)
-        ep_constraint_violation = np.asarray(self.env.queued_stats["constraint_violation"])
+        ep_constraint_violation = np.asarray(self.env.queued_stats['constraint_violation'])
         self.logger.add_scalars(
             {
-                "ep_length": ep_lengths.mean(),
-                "ep_return": ep_returns.mean(),
-                "ep_reward": (ep_returns / ep_lengths).mean(),
-                "ep_constraint_violation": ep_constraint_violation.mean()
+                'ep_length': ep_lengths.mean(),
+                'ep_return': ep_returns.mean(),
+                'ep_reward': (ep_returns / ep_lengths).mean(),
+                'ep_constraint_violation': ep_constraint_violation.mean()
             },
             step,
-            prefix="stat")
+            prefix='stat')
         # Total constraint violation during learning.
-        total_violations = self.env.accumulated_stats["constraint_violation"]
-        self.logger.add_scalars({"constraint_violation": total_violations}, step, prefix="stat")
-        if "eval" in results:
-            eval_ep_lengths = results["eval"]["ep_lengths"]
-            eval_ep_returns = results["eval"]["ep_returns"]
-            eval_constraint_violation = results["eval"]["constraint_violation"]
-            eval_mse = results["eval"]["mse"]
+        total_violations = self.env.accumulated_stats['constraint_violation']
+        self.logger.add_scalars({'constraint_violation': total_violations}, step, prefix='stat')
+        if 'eval' in results:
+            eval_ep_lengths = results['eval']['ep_lengths']
+            eval_ep_returns = results['eval']['ep_returns']
+            eval_constraint_violation = results['eval']['constraint_violation']
+            eval_mse = results['eval']['mse']
             self.logger.add_scalars(
                 {
-                    "ep_length": eval_ep_lengths.mean(),
-                    "ep_return": eval_ep_returns.mean(),
-                    "ep_reward": (eval_ep_returns / eval_ep_lengths).mean(),
-                    "constraint_violation": eval_constraint_violation.mean(),
-                    "mse": eval_mse.mean()
+                    'ep_length': eval_ep_lengths.mean(),
+                    'ep_return': eval_ep_returns.mean(),
+                    'ep_reward': (eval_ep_returns / eval_ep_lengths).mean(),
+                    'constraint_violation': eval_constraint_violation.mean(),
+                    'mse': eval_mse.mean()
                 },
                 step,
-                prefix="stat_eval")
+                prefix='stat_eval')
         # Print summary table
         self.logger.dump_scalars()
