@@ -1,16 +1,11 @@
-from collections import defaultdict
 from copy import deepcopy
+from collections import defaultdict
+
 import numpy as np
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
-from gymnasium.spaces import Box, Discrete
 
-from safe_control_gym.math_and_models.distributions import Normal, Categorical
-from safe_control_gym.math_and_models.neural_networks import MLP, CNN, RNN, init_
-from safe_control_gym.math_and_models.normalization import BaseNormalizer, MeanStdNormalizer, RewardStdNormalizer
-from safe_control_gym.math_and_models.schedule import *
-from safe_control_gym.math_and_models.random_processes import * 
+from safe_control_gym.math_and_models.neural_networks import MLP
 from safe_control_gym.controllers.sac.sac_utils import SACBuffer, soft_update
 
 
@@ -19,7 +14,7 @@ from safe_control_gym.controllers.sac.sac_utils import SACBuffer, soft_update
 # -----------------------------------------------------------------------------------
 
 class DDPGAgent:
-    """A DDPG class that encapsulates model, optimizer and update functions."""
+    '''A DDPG class that encapsulates model, optimizer and update functions.'''
 
     def __init__(self,
                  obs_space,
@@ -38,7 +33,7 @@ class DDPGAgent:
         self.tau = tau
 
         # model
-        self.ac = MLPActorCritic(obs_space, act_space, hidden_dims=[hidden_dim] * 2, activation="relu")
+        self.ac = MLPActorCritic(obs_space, act_space, hidden_dims=[hidden_dim] * 2, activation='relu')
 
         # target networks
         self.ac_targ = deepcopy(self.ac)
@@ -50,45 +45,45 @@ class DDPGAgent:
         self.critic_opt = torch.optim.Adam(self.ac.q.parameters(), critic_lr)
 
     def to(self, device):
-        """Puts agent to device."""
+        '''Puts agent to device.'''
         self.ac.to(device)
         self.ac_targ.to(device)
 
     def train(self):
-        """Sets training mode."""
+        '''Sets training mode.'''
         self.ac.train()
 
     def eval(self):
-        """Sets evaluation mode."""
+        '''Sets evaluation mode.'''
         self.ac.eval()
 
     def state_dict(self):
-        """Snapshots agent state."""
+        '''Snapshots agent state.'''
         return {
-            "ac": self.ac.state_dict(),
-            "ac_targ": self.ac_targ.state_dict(),
-            "actor_opt": self.actor_opt.state_dict(),
-            "critic_opt": self.critic_opt.state_dict(),
+            'ac': self.ac.state_dict(),
+            'ac_targ': self.ac_targ.state_dict(),
+            'actor_opt': self.actor_opt.state_dict(),
+            'critic_opt': self.critic_opt.state_dict(),
         }
 
     def load_state_dict(self, state_dict):
-        """Restores agent state."""
-        self.ac.load_state_dict(state_dict["ac"])
-        self.ac_targ.load_state_dict(state_dict["ac_targ"])
-        self.actor_opt.load_state_dict(state_dict["actor_opt"])
-        self.critic_opt.load_state_dict(state_dict["critic_opt"])
+        '''Restores agent state.'''
+        self.ac.load_state_dict(state_dict['ac'])
+        self.ac_targ.load_state_dict(state_dict['ac_targ'])
+        self.actor_opt.load_state_dict(state_dict['actor_opt'])
+        self.critic_opt.load_state_dict(state_dict['critic_opt'])
 
     def compute_policy_loss(self, batch):
-        """Returns policy loss(es) given batch of data."""
-        obs = batch["obs"]
+        '''Returns policy loss(es) given batch of data.'''
+        obs = batch['obs']
         act = self.ac.actor(obs)
         q = self.ac.q(obs, act)
         policy_loss = -q.mean()
         return policy_loss
 
     def compute_q_loss(self, batch):
-        """Returns q-value loss(es) given batch of data."""
-        obs, act, rew, next_obs, mask = batch["obs"], batch["act"], batch["rew"], batch["next_obs"], batch["mask"]
+        '''Returns q-value loss(es) given batch of data.'''
+        obs, act, rew, next_obs, mask = batch['obs'], batch['act'], batch['rew'], batch['next_obs'], batch['mask']
         q = self.ac.q(obs, act)
 
         with torch.no_grad():
@@ -101,7 +96,7 @@ class DDPGAgent:
         return critic_loss
 
     def update(self, batch):
-        """Updates model parameters based on current training batch."""
+        '''Updates model parameters based on current training batch.'''
         resutls = defaultdict(list)
 
         # actor update
@@ -119,8 +114,8 @@ class DDPGAgent:
         # update target networks
         soft_update(self.ac, self.ac_targ, self.tau)
 
-        resutls["policy_loss"] = policy_loss.item()
-        resutls["critic_loss"] = critic_loss.item()
+        resutls['policy_loss'] = policy_loss.item()
+        resutls['critic_loss'] = critic_loss.item()
         return resutls
 
 
@@ -154,9 +149,9 @@ class MLPQFunction(nn.Module):
 
 
 class MLPActorCritic(nn.Module):
-    """Model for the actor-critic agent."""
+    '''Model for the actor-critic agent.'''
 
-    def __init__(self, obs_space, act_space, hidden_dims=(64, 64), activation="relu"):
+    def __init__(self, obs_space, act_space, hidden_dims=(64, 64), activation='relu'):
         super().__init__()
         obs_dim = obs_space.shape[0]
         act_dim = act_space.shape[0]
@@ -182,14 +177,14 @@ class MLPActorCritic(nn.Module):
 # -----------------------------------------------------------------------------------
 
 class DDPGBuffer(SACBuffer):
-    """Storage for replay buffer during training.
+    '''Storage for replay buffer during training.
 
     Attributes:
         max_size (int): maximum size of the replay buffer.
         batch_size (int): number of samples (steps) per batch.
         scheme (dict): describs shape & other info of data to be stored.
         keys (list): names of all data from scheme.
-    """
+    '''
 
     def __init__(self, obs_space, act_space, max_size, batch_size=None):
         self.max_size = max_size
@@ -200,21 +195,21 @@ class DDPGBuffer(SACBuffer):
 
         N = max_size
         self.scheme = {
-            "obs": {
-                "vshape": (N, *obs_dim)
+            'obs': {
+                'vshape': (N, *obs_dim)
             },
-            "next_obs": {
-                "vshape": (N, *obs_dim)
+            'next_obs': {
+                'vshape': (N, *obs_dim)
             },
-            "act": {
-                "vshape": (N, act_dim)
+            'act': {
+                'vshape': (N, act_dim)
             },
-            "rew": {
-                "vshape": (N, 1)
+            'rew': {
+                'vshape': (N, 1)
             },
-            "mask": {
-                "vshape": (N, 1),
-                "init": np.ones
+            'mask': {
+                'vshape': (N, 1),
+                'init': np.ones
             }
         }
         self.keys = list(self.scheme.keys())
@@ -226,14 +221,13 @@ class DDPGBuffer(SACBuffer):
 # -----------------------------------------------------------------------------------
 
 def make_action_noise_process(noise_config, act_space):
-    """Construct a process for generating action noise during agent training.
-    """
-    process_func = noise_config.pop("func")
-    std_config = noise_config.pop("std")
+    '''Construct a process for generating action noise during agent training.'''
+    process_func = noise_config.pop('func')
+    std_config = noise_config.pop('std')
 
-    std_func = std_config.pop("func")
-    std_args = std_config.pop("args")
+    std_func = std_config.pop('func')
+    std_args = std_config.pop('args')
     std = eval(std_func)(*std_args, **std_config)
 
-    process = eval(process_func)(size=(act_space.shape[0],), std=std)     
+    process = eval(process_func)(size=(act_space.shape[0],), std=std)
     return process
