@@ -1,10 +1,9 @@
-"""Model Predictive Control.
-
-"""
-import numpy as np
-import casadi as cs
+'''Model Predictive Control.'''
 
 from copy import deepcopy
+
+import numpy as np
+import casadi as cs
 
 from safe_control_gym.controllers.base_controller import BaseController
 from safe_control_gym.controllers.mpc.mpc_utils import get_cost_weight_matrix, compute_discrete_lqr_gain_from_cont_linear_system, rk_discrete, compute_state_rmse, reset_constraints
@@ -13,9 +12,7 @@ from safe_control_gym.envs.constraints import GENERAL_CONSTRAINTS, create_constr
 
 
 class MPC(BaseController):
-    """MPC with full nonlinear model.
-
-    """
+    '''MPC with full nonlinear model.'''
 
     def __init__(
             self,
@@ -29,13 +26,13 @@ class MPC(BaseController):
             constraint_tol: float = 1e-6,
             # runner args
             # shared/base args
-            output_dir: str = "results/temp",
+            output_dir: str = 'results/temp',
             additional_constraints: list = None,
             use_gpu: bool = False,
             seed: int = 0,
             **kwargs
-            ):
-        """Creates task and controller.
+    ):
+        '''Creates task and controller.
 
         Args:
             env_func (Callable): function to instantiate task/environment.
@@ -50,14 +47,13 @@ class MPC(BaseController):
             additional_constraints (list): List of additional constraints
             use_gpu (bool): False (use cpu) True (use cuda).
             seed (int): random seed.
-
-        """
+        '''
         super().__init__(env_func, output_dir, use_gpu, seed, **kwargs)
         for k, v in locals().items():
-            if k != "self" and k != "kwargs" and "__" not in k:
+            if k != 'self' and k != 'kwargs' and '__' not in k:
                 self.__dict__.update({k: v})
 
-        #if prior_info is None:
+        # if prior_info is None:
         #    self.prior_info = {}
         # Task.
         self.env = env_func()
@@ -90,46 +86,40 @@ class MPC(BaseController):
     def add_constraints(self,
                         constraints
                         ):
-        """Add the constraints (from a list) to the system.
+        '''Add the constraints (from a list) to the system.
 
         Args:
             constraints (list): List of constraints controller is subject too.
-
-        """
+        '''
         self.constraints, self.state_constraints_sym, self.input_constraints_sym = reset_constraints(constraints + self.constraints.constraints)
 
     def remove_constraints(self,
                            constraints
                            ):
-        """Remove constraints from the current constraint list.
+        '''Remove constraints from the current constraint list.
 
         Args:
             constraints (list): list of constraints to be removed.
-
-        """
+        '''
         old_constraints_list = self.constraints.constraints
         for constraint in constraints:
             assert constraint in self.constraints.constraints,\
-                ValueError("This constraint is not in the current list of constraints")
+                ValueError('This constraint is not in the current list of constraints')
             old_constraints_list.remove(constraint)
         self.constraints, self.state_constraints_sym, self.input_constraints_sym = reset_constraints(old_constraints_list)
 
     def close(self):
-        """Cleans up resources.
-
-        """
+        '''Cleans up resources.'''
         self.env.close()
 
     def reset(self):
-        """Prepares for training or evaluation.
-
-        """
+        '''Prepares for training or evaluation.'''
         # Setup reference input.
         if self.env.TASK == Task.STABILIZATION:
-            self.mode = "stabilization"
+            self.mode = 'stabilization'
             self.x_goal = self.env.X_GOAL
         elif self.env.TASK == Task.TRAJ_TRACKING:
-            self.mode = "tracking"
+            self.mode = 'tracking'
             self.traj = self.env.X_GOAL.T
             # Step along the reference.
             self.traj_step = 0
@@ -144,10 +134,8 @@ class MPC(BaseController):
         self.setup_results_dict()
 
     def set_dynamics_func(self):
-        """Updates symbolic dynamics with actual control frequency.
-
-        """
-        #self.dynamics_func = cs.integrator('fd', 'rk',
+        '''Updates symbolic dynamics with actual control frequency.'''
+        # self.dynamics_func = cs.integrator('fd', 'rk',
         #                                   {
         #                                    'x': self.model.x_sym,
         #                                    'p': self.model.u_sym,
@@ -160,29 +148,25 @@ class MPC(BaseController):
                                          self.dt)
 
     def compute_initial_guess(self, init_state, goal_states, x_lin, u_lin):
-        """Use LQR to get an initial guess of the """
+        '''Use LQR to get an initial guess of the '''
         dfdxdfdu = self.model.df_func(x=x_lin, u=u_lin)
         dfdx = dfdxdfdu['dfdx'].toarray()
         dfdu = dfdxdfdu['dfdu'].toarray()
-        lqr_gain, Ad, Bd = compute_discrete_lqr_gain_from_cont_linear_system(dfdx, dfdu, self.Q, self.R, self.dt)
+        lqr_gain, _, _ = compute_discrete_lqr_gain_from_cont_linear_system(dfdx, dfdu, self.Q, self.R, self.dt)
 
-        x_guess = np.zeros((self.model.nx, self.T+1))
+        x_guess = np.zeros((self.model.nx, self.T + 1))
         u_guess = np.zeros((self.model.nu, self.T))
         x_guess[:, 0] = init_state
 
         for i in range(self.T):
             u = lqr_gain @ (x_guess[:, i] - goal_states[:, i]) + u_lin
             u_guess[:, i] = u
-            x_guess[:, i+1, None] = self.dynamics_func(x0=x_guess[:, i], p=u)['xf'].toarray()
+            x_guess[:, i + 1, None] = self.dynamics_func(x0=x_guess[:, i], p=u)['xf'].toarray()
 
         return x_guess, u_guess
 
-
-
     def setup_optimizer(self):
-        """Sets up nonlinear optimization problem.
-
-        """
+        '''Sets up nonlinear optimization problem.'''
         nx, nu = self.model.nx, self.model.nu
         T = self.T
         # Define optimizer and variables.
@@ -209,14 +193,14 @@ class MPC(BaseController):
                               Xr=x_ref[:, i],
                               Ur=np.zeros((nu, 1)),
                               Q=self.Q,
-                              R=self.R)["l"]
+                              R=self.R)['l']
         # Terminal cost.
         cost += cost_func(x=x_var[:, -1],
                           u=np.zeros((nu, 1)),
                           Xr=x_ref[:, -1],
                           Ur=np.zeros((nu, 1)),
                           Q=self.Q,
-                          R=self.R)["l"]
+                          R=self.R)['l']
         # Constraints
         for i in range(self.T):
             # Dynamics constraints.
@@ -226,17 +210,17 @@ class MPC(BaseController):
             for sc_i, state_constraint in enumerate(self.state_constraints_sym):
                 if self.soft_constraints:
                     opti.subject_to(state_constraint(x_var[:, i]) <= state_slack[sc_i])
-                    cost += 10000*state_slack[sc_i]**2
+                    cost += 10000 * state_slack[sc_i]**2
                     opti.subject_to(state_slack[sc_i] >= 0)
                 else:
                     opti.subject_to(state_constraint(x_var[:, i]) < -self.constraint_tol)
             for ic_i, input_constraint in enumerate(self.input_constraints_sym):
                 if self.soft_constraints:
-                    opti.subject_to(input_constraint(u_var[:,i]) <= input_slack[ic_i])
-                    cost += 10000*input_slack[ic_i]**2
+                    opti.subject_to(input_constraint(u_var[:, i]) <= input_slack[ic_i])
+                    cost += 10000 * input_slack[ic_i]**2
                     opti.subject_to(input_slack[ic_i] >= 0)
                 else:
-                    opti.subject_to(input_constraint(u_var[:,i]) < -self.constraint_tol)
+                    opti.subject_to(input_constraint(u_var[:, i]) < -self.constraint_tol)
 
         # Final state constraints.
         for sc_i, state_constraint in enumerate(self.state_constraints_sym):
@@ -251,23 +235,23 @@ class MPC(BaseController):
 
         opti.minimize(cost)
         # Create solver (IPOPT solver in this version)
-        #opts = {"ipopt.print_level": 0, "ipopt.sb": "yes", "print_time": 0}
-        opts = {"expand": True}
+        #opts = {'ipopt.print_level': 0, 'ipopt.sb': 'yes', 'print_time': 0}
+        opts = {'expand': True}
         opti.solver('ipopt', opts)
         self.opti_dict = {
-            "opti": opti,
-            "x_var": x_var,
-            "u_var": u_var,
-            "x_init": x_init,
-            "x_ref": x_ref,
-            "cost": cost
+            'opti': opti,
+            'x_var': x_var,
+            'u_var': u_var,
+            'x_init': x_init,
+            'x_ref': x_ref,
+            'cost': cost
         }
 
     def select_action(self,
                       obs,
                       info=None
                       ):
-        """Solves nonlinear mpc problem to get next action.
+        '''Solves nonlinear mpc problem to get next action.
 
         Args:
             obs (ndarray): Current state/observation.
@@ -275,27 +259,27 @@ class MPC(BaseController):
 
         Returns:
             action (ndarray): Input/action to the task/env.
-        """
+        '''
 
         opti_dict = self.opti_dict
-        opti = opti_dict["opti"]
-        x_var = opti_dict["x_var"]
-        u_var = opti_dict["u_var"]
-        x_init = opti_dict["x_init"]
-        x_ref = opti_dict["x_ref"]
-        cost = opti_dict["cost"]
+        opti = opti_dict['opti']
+        x_var = opti_dict['x_var']
+        u_var = opti_dict['u_var']
+        x_init = opti_dict['x_init']
+        x_ref = opti_dict['x_ref']
+
         # Assign the initial state.
         opti.set_value(x_init, obs)
         # Assign reference trajectory within horizon.
         goal_states = self.get_references()
         opti.set_value(x_ref, goal_states)
-        if self.mode == "tracking":
+        if self.mode == 'tracking':
             self.traj_step += 1
-        #if self.warmstart and self.x_prev is None and self.u_prev is None:
+        # if self.warmstart and self.x_prev is None and self.u_prev is None:
         #    x_guess, u_guess = self.compute_initial_guess(obs, goal_states, self.X_EQ, self.U_EQ)
         #    opti.set_initial(x_var, x_guess)
         #    opti.set_initial(u_var, u_guess) # Initial guess for optimization problem.
-        #elif self.warmstart and self.x_prev is not None and self.u_prev is not None:
+        # elif self.warmstart and self.x_prev is not None and self.u_prev is not None:
         if self.warmstart and self.x_prev is not None and self.u_prev is not None:
             # shift previous solutions by 1 step
             x_guess = deepcopy(self.x_prev)
@@ -322,9 +306,7 @@ class MPC(BaseController):
         return action
 
     def get_references(self):
-        """Constructs reference states along mpc horizon.(nx, T+1).
-
-        """
+        '''Constructs reference states along mpc horizon.(nx, T+1).'''
         if self.env.TASK == Task.STABILIZATION:
             # Repeat goal state for horizon steps.
             goal_states = np.tile(self.env.X_GOAL.reshape(-1, 1), (1, self.T + 1))
@@ -338,28 +320,26 @@ class MPC(BaseController):
                 np.tile(self.traj[:, -1:], (1, remain))
             ], -1)
         else:
-            raise Exception("Reference for this mode is not implemented.")
+            raise Exception('Reference for this mode is not implemented.')
         return goal_states  # (nx, T+1).
 
     def setup_results_dict(self):
-        """
-
-        """
-        self.results_dict = { 'obs': [],
-                              'reward': [],
-                              'done': [],
-                              'info': [],
-                              'action': [],
-                              'horizon_inputs': [],
-                              'horizon_states': [],
-                              'goal_states': [],
-                              'frames': [],
-                              'state_mse': [],
-                              'common_cost': [],
-                              'state': [],
-                              'state_error': [],
-                              't_wall': []
-        }
+        '''Setup the results dictionary to store run information.'''
+        self.results_dict = {'obs': [],
+                             'reward': [],
+                             'done': [],
+                             'info': [],
+                             'action': [],
+                             'horizon_inputs': [],
+                             'horizon_states': [],
+                             'goal_states': [],
+                             'frames': [],
+                             'state_mse': [],
+                             'common_cost': [],
+                             'state': [],
+                             'state_error': [],
+                             't_wall': []
+                             }
 
     def run(self,
             env=None,
@@ -368,7 +348,7 @@ class MPC(BaseController):
             max_steps=None,
             terminate_run_on_done=None
             ):
-        """Runs evaluation with current policy.
+        '''Runs evaluation with current policy.
 
         Args:
             render (bool): if to do real-time rendering.
@@ -376,8 +356,7 @@ class MPC(BaseController):
 
         Returns:
             dict: evaluation statisitcs, rendered frames.
-
-        """
+        '''
         if env is None:
             env = self.env
         if terminate_run_on_done is None:
@@ -389,7 +368,7 @@ class MPC(BaseController):
             env.set_cost_function_param(self.Q, self.R)
         #obs, info = env.reset()
         obs = env.reset()
-        print("Init State:")
+        print('Init State:')
         print(obs)
         ep_returns, ep_lengths = [], []
         frames = []
@@ -399,7 +378,7 @@ class MPC(BaseController):
         i = 0
         if env.TASK == Task.STABILIZATION:
             if max_steps is None:
-                MAX_STEPS = int(env.CTRL_FREQ*env.EPISODE_LEN_SEC)
+                MAX_STEPS = int(env.CTRL_FREQ * env.EPISODE_LEN_SEC)
             else:
                 MAX_STEPS = max_steps
         elif env.TASK == Task.TRAJ_TRACKING:
@@ -408,14 +387,14 @@ class MPC(BaseController):
             else:
                 MAX_STEPS = max_steps
         else:
-            raise("Undefined Task")
+            raise Exception('Undefined Task')
         self.terminate_loop = False
         done = False
         common_metric = 0
-        while not(done and terminate_run_on_done) and i < MAX_STEPS and not (self.terminate_loop):
+        while not (done and terminate_run_on_done) and i < MAX_STEPS and not (self.terminate_loop):
             action = self.select_action(obs)
             if self.terminate_loop:
-                print("Infeasible MPC Problem")
+                print('Infeasible MPC Problem')
                 break
             # Repeat input for more efficient control.
             obs, reward, done, info = env.step(action)
@@ -425,10 +404,10 @@ class MPC(BaseController):
             self.results_dict['info'].append(info)
             self.results_dict['action'].append(action)
             self.results_dict['state'].append(env.state)
-            self.results_dict['state_mse'].append(info["mse"])
+            self.results_dict['state_mse'].append(info['mse'])
             #self.results_dict['state_error'].append(env.state - env.X_GOAL[i,:])
 
-            common_metric += info["mse"]
+            common_metric += info['mse']
             print(i, '-th step.')
             print(action)
             print(obs)
@@ -438,14 +417,14 @@ class MPC(BaseController):
             print()
             if render:
                 env.render()
-                frames.append(env.render("rgb_array"))
+                frames.append(env.render('rgb_array'))
             i += 1
         # Collect evaluation results.
         ep_lengths = np.asarray(ep_lengths)
         ep_returns = np.asarray(ep_returns)
         if logging:
-            msg = "****** Evaluation ******\n"
-            msg += "eval_ep_length {:.2f} +/- {:.2f} | eval_ep_return {:.3f} +/- {:.3f}\n".format(
+            msg = '****** Evaluation ******\n'
+            msg += 'eval_ep_length {:.2f} +/- {:.2f} | eval_ep_return {:.3f} +/- {:.3f}\n'.format(
                 ep_lengths.mean(), ep_lengths.std(), ep_returns.mean(),
                 ep_returns.std())
         if len(frames) != 0:
@@ -459,9 +438,16 @@ class MPC(BaseController):
             self.results_dict['total_rmse_state_error'] = compute_state_rmse(self.results_dict['state'])
             self.results_dict['total_rmse_obs_error'] = compute_state_rmse(self.results_dict['obs'])
         except ValueError:
-            raise Exception("[ERROR] mpc.run().py: MPC could not find a solution for the first step given the initial conditions. "
-                  "Check to make sure initial conditions are feasible.")
+            raise Exception('[ERROR] mpc.run().py: MPC could not find a solution for the first step given the initial conditions. '
+                            'Check to make sure initial conditions are feasible.')
         return deepcopy(self.results_dict)
 
     def reset_before_run(self, obs, info=None, env=None):
+        '''Reinitialize just the controller before a new run.
+
+        Args:
+            obs (ndarray): The initial observation for the new run.
+            info (dict): The first info of the new run.
+            env (BenchmarkEnv): The environment to be used for the new run.
+        '''
         self.reset()
