@@ -1,6 +1,9 @@
-import numpy as np
+'''Initializing vectorized environments.'''
+
 import random
+
 import torch
+import numpy as np
 
 from safe_control_gym.envs.env_wrappers.vectorized_env.dummy_vec_env import DummyVecEnv
 from safe_control_gym.envs.env_wrappers.vectorized_env.subproc_vec_env import SubprocVecEnv
@@ -10,18 +13,17 @@ def make_env_fn(env_func,
                 env_config,
                 seed,
                 rank):
-    """Higher-order function for env init func.
+    '''Higher-order function for env init func.
 
     Args:
         env_func (function): partial function that can accept args.
         env_config (dict): env init args.
         seed (int): base seed.
         rank (int): unique seed for each parallel env.
-        
-    Returns:
-        _thunk: env-constructing func.
 
-    """
+    Returns:
+        _thunk (func): env-constructing func.
+    '''
 
     def _thunk():
         # Do not set seed i if 0 (e.g. for evaluation).
@@ -42,19 +44,18 @@ def make_vec_envs(env_func,
                   batch_size=1,
                   n_processes=1,
                   seed=None):
-    """Produce envs with parallel rollout abilities. 
+    '''Produce envs with parallel rollout abilities.
 
     Args:
         env_func (function): partial function that can accept args.
-        env_config (dict): non-shareable args for each env. 
-        batch_size (int): total num of parallel envs. 
+        env_config (dict): non-shareable args for each env.
+        batch_size (int): total num of parallel envs.
         n_processes (int): num of parallel workers to run envs.
-        seed (int): base seed for the run. 
+        seed (int): base seed for the run.
 
     Returns:
-        VecEnv: (wrapped) parallel envs.
-
-    """
+        VecEnv (Vectorized env): (wrapped) parallel envs.
+    '''
     if env_configs is None:
         env_configs = [{}] * batch_size
     env_fns = [make_env_fn(env_func, env_configs[i], seed, i) for i in range(batch_size)]
