@@ -1,12 +1,10 @@
-"""Symbolic Models.
+'''Symbolic Models.'''
 
-"""
-import numpy as np
 import casadi as cs
 
 
 class SymbolicModel():
-    """Implements the dynamics model with symbolic variables.
+    '''Implements the dynamics model with symbolic variables.
 
     x_dot = f(x,u), y = g(x,u), with other pre-defined, symbolic functions
     (e.g. cost, constraints), serve as priors for the controllers.
@@ -15,8 +13,7 @@ class SymbolicModel():
         * naming convention on symbolic variable and functions.
             * for single-letter symbol, use {}_sym, otherwise use underscore for delimiter.
             * for symbolic functions to be exposed, use {}_func.
-
-    """
+    '''
 
     def __init__(self,
                  dynamics,
@@ -26,17 +23,14 @@ class SymbolicModel():
                  funcs=None,
                  params=None,
                  ):
-        """
-
-        """
         # Setup for dynamics.
-        self.x_sym = dynamics["vars"]["X"]
-        self.u_sym = dynamics["vars"]["U"]
-        self.x_dot = dynamics["dyn_eqn"]
-        if dynamics["obs_eqn"] is None:
+        self.x_sym = dynamics['vars']['X']
+        self.u_sym = dynamics['vars']['U']
+        self.x_dot = dynamics['dyn_eqn']
+        if dynamics['obs_eqn'] is None:
             self.y_sym = self.x_sym
         else:
-            self.y_sym = dynamics["obs_eqn"]
+            self.y_sym = dynamics['obs_eqn']
         # Sampling time.
         self.dt = dt
         # Integration algorithm.
@@ -57,35 +51,31 @@ class SymbolicModel():
         self.nu = self.u_sym.shape[0]
         self.ny = self.y_sym.shape[0]
         # Setup cost function.
-        self.cost_func = cost["cost_func"]
+        self.cost_func = cost['cost_func']
         # print(self.cost_func)
-        self.Q = cost["vars"]["Q"]
-        self.R = cost["vars"]["R"]
-        self.Xr = cost["vars"]["Xr"]
-        self.Ur = cost["vars"]["Ur"]
+        self.Q = cost['vars']['Q']
+        self.R = cost['vars']['R']
+        self.Xr = cost['vars']['Xr']
+        self.Ur = cost['vars']['Ur']
         # Setup symbolic model.
         self.setup_model()
         # Setup Jacobian and Hessian of the dynamics and cost functions.
         self.setup_linearization()
 
     def setup_model(self):
-        """Exposes functions to evaluate the model.
-
-        """
+        '''Exposes functions to evaluate the model.'''
         # Continuous time dynamics.
         self.fc_func = cs.Function('fc', [self.x_sym, self.u_sym], [self.x_dot], ['x', 'u'], ['f'])
         # Discrete time dynamics.
         self.fd_func = cs.integrator('fd', self.integration_algo, {'x': self.x_sym,
                                                                    'p': self.u_sym,
                                                                    'ode': self.x_dot}, {'tf': self.dt}
-                                    )
+                                     )
         # Observation model.
         self.g_func = cs.Function('g', [self.x_sym, self.u_sym], [self.y_sym], ['x', 'u'], ['g'])
 
     def setup_linearization(self):
-        """Exposes functions for the linearized model.
-
-        """
+        '''Exposes functions for the linearized model.'''
         # Jacobians w.r.t state & input.
         self.dfdx = cs.jacobian(self.x_dot, self.x_sym)
         self.dfdu = cs.jacobian(self.x_dot, self.u_sym)
