@@ -1,20 +1,20 @@
-'''Register safety filters. '''
+'''Register safety filters.'''
 
 from safe_control_gym.utils.registration import register
 
 
-register(id='linear_mpsc',
+register(idx='linear_mpsc',
          entry_point='safe_control_gym.safety_filters.mpsc.linear_mpsc:LINEAR_MPSC',
          config_entry_point='safe_control_gym.safety_filters.mpsc:mpsc.yaml')
 
-register(id='nl_mpsc',
+register(idx='nl_mpsc',
          entry_point='safe_control_gym.safety_filters.mpsc.nl_mpsc:NL_MPSC',
          config_entry_point='safe_control_gym.safety_filters.mpsc:mpsc.yaml')
 
-register(id='cbf',
+register(idx='cbf',
          entry_point='safe_control_gym.safety_filters.cbf.cbf:CBF',
          config_entry_point='safe_control_gym.safety_filters.cbf:cbf.yaml')
 
-register(id='cbf_nn',
+register(idx='cbf_nn',
          entry_point='safe_control_gym.safety_filters.cbf.cbf_nn:CBF_NN',
          config_entry_point='safe_control_gym.safety_filters.cbf:cbf_nn.yaml')
