@@ -1,10 +1,10 @@
-'''A cost assuming a constant control input. '''
+'''A cost assuming a constant control input.'''
 
 from safe_control_gym.safety_filters.mpsc.mpsc_cost_function.abstract_cost import MPSC_COST
 
 
 class CONSTANT_COST(MPSC_COST):
-    '''Constant Cost. '''
+    '''Constant Cost.'''
 
     def get_cost(self, opti_dict):
         '''Returns the cost function for the MPSC optimization in symbolic form.
@@ -22,5 +22,5 @@ class CONSTANT_COST(MPSC_COST):
 
         cost = (u_L - next_u).T @ (u_L - next_u)
         for h in range(1, self.mpsc_cost_horizon):
-            cost += (self.decay_factor**h)*(u_L - v_var[:, h]).T @ (u_L - v_var[:, h])
+            cost += (self.decay_factor**h) * (u_L - v_var[:, h]).T @ (u_L - v_var[:, h])
         return cost
